remove old type visitor and move in struct visitor
